Aula 7/tarefa 3/requirements.txt
requests
folium
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv